Move to `chia/wallet/util/pprint.py`
